refactor: テーブルページをcommonフォルダに移動
