+++ frontend/src/components/pages/ManagerPage.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import GenericTemplate from '../templates/GenericTemplate';
 import axios from 'axios';
 import _ from 'lodash';
-import TablePages, { HeadCell } from './TablePages';
+import TablePages, { HeadCell } from '../common/TablePages';
 
 const THREE = '(3年以上)';
 const ALL = '(全て)';
