+++ frontend/src/components/pages/PlayerPage.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { RouteComponentProps } from 'react-router-dom';
 import GenericTemplate from '../templates/GenericTemplate';
 import Button from '@material-ui/core/Button';
-import TablePages, { HeadCell } from './TablePages';
+import TablePages, { HeadCell } from '../common/TablePages';
 import axios from 'axios';
 import _ from 'lodash';
